Move combinator methods to the `Body` trait
Not sure it makes sense to have them on a separate trait when everything
is local to the crate.
