+++ src/combinators/map_err.rs
